pyproject.toml
[project]
name = "logs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bytewax>=0.21.0",
    "fastapi>=0.115.4",
    "numpy>=2.1.3",
    "pydantic>=2.9.2",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "alembic>=1.14.0",
    "httpx>=0.27.2",
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "ruff>=0.7.2",
    "types-python-dateutil>=2.9.0.20241003",
]

[tool.ruff]
line-length = 120

[tool.mypy]
files = ["uptime"]
